[ci] Change comment on what empty constructor exists
maven doesn't use plexus any longer so that information was incorrect.  The reason reason beyond that is that java has deprecated default constructors in general due to fact that any time one is overridden it disappears, generally speaking, its better to define empty one (or use some default in it) when overriding the default behavior so lean on this being more for java need than what maven might be doing (maven uses jsr 330 now which does want empty constructor too).